Create new selection test file and copy paste tests
