+++ extensions/ql-vscode/test/vscode-tests/minimal-workspace/databases/db-panel.test.ts
@@ -9,10 +9,7 @@ import { DbManager } from "../../../../src/databases/db-manager";
 import { DbConfigStore } from "../../../../src/databases/config/db-config-store";
 import { DbTreeDataProvider } from "../../../../src/databases/ui/db-tree-data-provider";
 import { DbItemKind, DbListKind } from "../../../../src/databases/db-item";
-import {
-  DbTreeViewItem,
-  SELECTED_DB_ITEM_RESOURCE_URI,
-} from "../../../../src/databases/ui/db-tree-view-item";
+import { DbTreeViewItem } from "../../../../src/databases/ui/db-tree-view-item";
 import { ExtensionApp } from "../../../../src/common/vscode/vscode-app";
 import { createMockExtensionContext } from "../../../factories/extension-context";
 import { createDbConfig } from "../../../factories/db-config-factories";
@@ -49,113 +46,6 @@ describe("db panel", () => {
 
   afterEach(async () => {
     await remove(workspaceStoragePath);
-  });
-
-  describe("selecting an item", () => {
-    it("should mark selected remote db list as selected", async () => {
-      const dbConfig: DbConfig = createDbConfig({
-        remoteLists: [
-          {
-            name: "my-list-1",
-            repositories: ["owner1/repo1", "owner1/repo2"],
-          },
-          {
-            name: "my-list-2",
-            repositories: ["owner2/repo1", "owner2/repo2"],
-          },
-        ],
-        selected: {
-          kind: SelectedDbItemKind.VariantAnalysisUserDefinedList,
-          listName: "my-list-2",
-        },
-      });
-
-      await saveDbConfig(dbConfig);
-
-      const dbTreeItems = await dbTreeDataProvider.getChildren();
-
-      expect(dbTreeItems).toBeTruthy();
-      const items = dbTreeItems!;
-
-      const remoteRootNode = items[0];
-      expect(remoteRootNode.dbItem).toBeTruthy();
-      expect(remoteRootNode.dbItem?.kind).toEqual(DbItemKind.RootRemote);
-
-      const list1 = remoteRootNode.children.find(
-        (c) =>
-          c.dbItem?.kind === DbItemKind.VariantAnalysisUserDefinedList &&
-          c.dbItem?.listName === "my-list-1",
-      );
-      const list2 = remoteRootNode.children.find(
-        (c) =>
-          c.dbItem?.kind === DbItemKind.VariantAnalysisUserDefinedList &&
-          c.dbItem?.listName === "my-list-2",
-      );
-
-      expect(list1).toBeTruthy();
-      expect(list2).toBeTruthy();
-      expect(isTreeViewItemSelectable(list1!)).toBeTruthy();
-      expect(isTreeViewItemSelected(list2!)).toBeTruthy();
-    });
-
-    it("should mark selected remote db inside list as selected", async () => {
-      const dbConfig: DbConfig = createDbConfig({
-        remoteLists: [
-          {
-            name: "my-list-1",
-            repositories: ["owner1/repo1", "owner1/repo2"],
-          },
-          {
-            name: "my-list-2",
-            repositories: ["owner1/repo1", "owner2/repo2"],
-          },
-        ],
-        remoteRepos: ["owner1/repo1"],
-        selected: {
-          kind: SelectedDbItemKind.VariantAnalysisRepository,
-          repositoryName: "owner1/repo1",
-          listName: "my-list-2",
-        },
-      });
-
-      await saveDbConfig(dbConfig);
-
-      const dbTreeItems = await dbTreeDataProvider.getChildren();
-
-      expect(dbTreeItems).toBeTruthy();
-      const items = dbTreeItems!;
-
-      const remoteRootNode = items[0];
-      expect(remoteRootNode.dbItem).toBeTruthy();
-      expect(remoteRootNode.dbItem?.kind).toEqual(DbItemKind.RootRemote);
-
-      const list2 = remoteRootNode.children.find(
-        (c) =>
-          c.dbItem?.kind === DbItemKind.VariantAnalysisUserDefinedList &&
-          c.dbItem?.listName === "my-list-2",
-      );
-      expect(list2).toBeTruthy();
-
-      const repo1Node = list2?.children.find(
-        (c) =>
-          c.dbItem?.kind === DbItemKind.RemoteRepo &&
-          c.dbItem?.repoFullName === "owner1/repo1",
-      );
-      expect(repo1Node).toBeTruthy();
-      expect(isTreeViewItemSelected(repo1Node!)).toBeTruthy();
-
-      const repo2Node = list2?.children.find(
-        (c) =>
-          c.dbItem?.kind === DbItemKind.RemoteRepo &&
-          c.dbItem?.repoFullName === "owner2/repo2",
-      );
-      expect(repo2Node).toBeTruthy();
-      expect(isTreeViewItemSelectable(repo2Node!)).toBeTruthy();
-
-      for (const item of remoteRootNode.children) {
-        expect(isTreeViewItemSelectable(item)).toBeTruthy();
-      }
-    });
   });
 
   describe("addNewRemoteRepo", () => {
@@ -496,22 +386,6 @@ describe("db panel", () => {
     expect(item.tooltip).toBe(tooltip);
     expect(item.collapsibleState).toBe(TreeItemCollapsibleState.None);
     expect(item.children.length).toBe(0);
-  }
-
-  function isTreeViewItemSelectable(treeViewItem: DbTreeViewItem) {
-    return (
-      treeViewItem.resourceUri === undefined &&
-      treeViewItem.contextValue?.includes("canBeSelected")
-    );
-  }
-
-  function isTreeViewItemSelected(treeViewItem: DbTreeViewItem) {
-    return (
-      treeViewItem.resourceUri?.toString(true) ===
-        SELECTED_DB_ITEM_RESOURCE_URI &&
-      (treeViewItem.contextValue === undefined ||
-        !treeViewItem.contextValue.includes("canBeSelected"))
-    );
   }
 
   async function readDbConfigDirectly(): Promise<DbConfig> {
